current codes

+++ my_robot_streamer/my_robot_streamer/bridge_node.py
@@ -5,6 +5,8 @@ from cv_bridge import CvBridge
 import paho.mqtt.client as mqtt
 import cv2
 import yaml
+import os
+import base64
 
 class RosMqttVideoBridge(Node):
     def __init__(self, mqtt_config_path):
@@ -16,12 +18,12 @@ class RosMqttVideoBridge(Node):
 
         self.subscription = self.create_subscription(
             Image,
-            '/image_raw',  # Replace with your ROS image topic
+            '/image_raw',
             self.callback,
             10  # Message queue size
         )
 
-        self.mqtt_client = mqtt.Client()
+        self.mqtt_client = mqtt.Client(mqtt.CallbackAPIVersion.VERSION2)
         self.mqtt_client.on_connect = self.on_connect
         self.mqtt_client.on_disconnect = self.on_disconnect
 
@@ -36,18 +38,20 @@ class RosMqttVideoBridge(Node):
 
     def on_connect(self, client, userdata, flags, rc):
         self.get_logger().info("Connected to MQTT broker with result code %s" % rc)
+        print("Connected to MQTT broker with result code %s" % rc)
 
     def on_disconnect(self, client, userdata, rc):
         self.get_logger().info("Disconnected from MQTT broker with result code %s" % rc)
+        print("Disconnected from MQTT broker with result code %s" % rc)
 
     def callback(self, msg):
         try:
             bridge = CvBridge()
-            cv_image = bridge.imgmsg_to_cv2(msg, desired_encoding='bgr8')
+            cv_image = bridge.imgmsg_to_cv2(msg,'bgr8')
             _, image_data = cv2.imencode('.jpg', cv_image)
-            payload = image_data.tobytes()
+            payload = base64.b64encode(image_data)
 
-            mqtt_topic = "/ros/mqtt_video"  # Replace with your MQTT topic for video frames
+            mqtt_topic = "mqtt_video"
 
             self.mqtt_client.publish(mqtt_topic, payload, qos=0, retain=False)
             self.get_logger().info(f"Published video frame to MQTT")
@@ -58,13 +62,16 @@ class RosMqttVideoBridge(Node):
 def main(args=None):
     rclpy.init(args=args)
 
-    # Provide the path to your MQTT configuration YAML file
     mqtt_config_path = '/home/nikhil/Desktop/ros2_ws/src/my_robot_streamer/config/ros_mqtt_config.yaml'
+    if os.path.exists(mqtt_config_path):
 
-    node = RosMqttVideoBridge(mqtt_config_path)
-    rclpy.spin(node)
-    node.destroy_node()
-    rclpy.shutdown()
+        node = RosMqttVideoBridge(mqtt_config_path)
+
+        rclpy.spin(node)
+        node.destroy_node()
+        rclpy.shutdown()
+    else:
+        print(f"{mqtt_config_path},file path doesnt exist")
 
 if __name__ == '__main__':
     main()
